refactor test 전용 클래스 test 패키지로 리팩토링
