+++ src/test/java/model/RacingGameTest.java
@@ -3,6 +3,8 @@ package model;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import test.FixedNumberGenerator;
+
 import static org.assertj.core.api.Assertions.*;
 import java.util.List;
